fix(tcp): accumulate congestion events per-interval for reliable markers
Events (zero-window, dup-ack, retransmit, ECE, CWR) were being queried
with timestamp lookback at publication time, causing markers to appear
at wrong positions or not at all. Now events are accumulated per-interval
during packet processing, matching how bytes/packets work.

Changes:
- tcp_window_process_packet() returns detected events as uint64_t bitmask
- propagate_events_to_intervals() ORs events into all interval tables
- Edge-triggered zero-window detection with 5% hysteresis threshold
- Frontend simplified to use backend-provided per-interval events
- Screenshot controller updated to capture RTT/Window chart containers
  with titles and legends

+++ deps/toptalk/flow.h
@@ -176,7 +176,7 @@ struct flow_window_info {
 	uint32_t dup_ack_cnt;     /* Duplicate ACK events */
 	uint32_t retransmit_cnt;  /* Retransmission events */
 	uint32_t ece_cnt;         /* ECE flag count */
-	uint8_t recent_events;    /* Bitmask of recent congestion events */
+	uint64_t recent_events;   /* Bitmask of recent congestion events */
 };
 
 /* Inter-packet gap (IPG) histogram bucket count */
